Move auth implementations to cpp
