+++ include/ccf/endpoint_context.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/endpoints/authentication/authentication_types.h"
 #include "enclave/rpc_context.h"
-#include "http/authentication/authentication_types.h"
 
 #include <functional>
 #include <memory>
